Combine integration test binaries
Each .rs file in tests translates to a separate binary, resulting in
longer build and run times than we could achieve by combining tests.

https://doc.rust-lang.org/rust-by-example/testing/integration_testing.html
