+++ tests/common/mod.rs
@@ -1,0 +1,21 @@
+#![allow(dead_code)]
+
+pub mod delegate;
+pub mod features;
+pub mod options;
+pub mod palworld;
+
+pub const ASSERT_FAILED_PATH: &str = "resources/test/assert_failed.sav";
+pub const COMPONENT8_PATH: &str = "resources/test/component8.sav";
+pub const DELEGATE_PATH: &str = "resources/test/Delegate.sav";
+pub const ENUM_ARRAY_PATH: &str = "resources/test/enum_array.sav";
+pub const FEATURES_01_PATH: &str = "resources/test/features_01.bin";
+pub const OPTIONS_PATH: &str = "resources/test/Options.sav";
+pub const PACKAGE_VERSION_524_PATH: &str = "resources/test/package_version_524.sav";
+pub const PALWORLD_ZLIB_PATH: &str = "resources/test/palworld_zlib.sav";
+pub const PALWORLD_ZLIB_TWICE_PATH: &str = "resources/test/palworld_zlib_twice.sav";
+pub const REGRESSION_01_PATH: &str = "resources/test/regression_01.bin";
+pub const SLOT1_PATH: &str = "resources/test/Slot1.sav";
+pub const SLOT2_PATH: &str = "resources/test/Slot2.sav";
+pub const SLOT3_PATH: &str = "resources/test/Slot3.sav";
+pub const TRANSFORM_PATH: &str = "resources/test/transform.sav";
