+++ tests/common/palworld.rs
@@ -1,0 +1,80 @@
+use std::collections::HashMap;
+
+pub(crate) fn hints() -> HashMap<String, String> {
+    HashMap::from([(
+        String::from(
+            "worldSaveData.StructProperty.CharacterSaveParameterMap.MapProperty.Key.StructProperty",
+        ),
+        String::from("StructProperty"),
+    ), (
+        String::from(
+            "worldSaveData.StructProperty.CharacterSaveParameterMap.MapProperty.Value.StructProperty",
+        ),
+        String::from("StructProperty"),
+    ), (
+        String::from(
+            "worldSaveData.StructProperty.MapObjectSaveData.ArrayProperty.ConcreteModel.StructProperty.ModuleMap.MapProperty.Value.StructProperty",
+        ),
+        String::from("StructProperty"),
+    ), (
+        String::from("worldSaveData.StructProperty.FoliageGridSaveDataMap.MapProperty.Key.StructProperty"),
+        String::from("StructProperty"),
+    ), (
+        String::from("worldSaveData.StructProperty.FoliageGridSaveDataMap.MapProperty.Value.StructProperty"),
+        String::from("StructProperty"),
+    ), (
+        String::from("worldSaveData.StructProperty.FoliageGridSaveDataMap.MapProperty.Value.StructProperty.ModelMap.MapProperty.Value.StructProperty"),
+        String::from("StructProperty")
+    ), (
+        String::from("worldSaveData.StructProperty.FoliageGridSaveDataMap.MapProperty.Value.StructProperty.ModelMap.MapProperty.Value.StructProperty.InstanceDataMap.MapProperty.Key.StructProperty"),
+        String::from("StructProperty")
+    ), (
+        String::from("worldSaveData.StructProperty.FoliageGridSaveDataMap.MapProperty.Value.StructProperty.ModelMap.MapProperty.Value.StructProperty.InstanceDataMap.MapProperty.Value.StructProperty"),
+        String::from("StructProperty")
+    ), (
+        String::from("worldSaveData.StructProperty.MapObjectSpawnerInStageSaveData.MapProperty.Key.StructProperty"), 
+        String::from("StructProperty")
+    ), (
+        String::from("worldSaveData.StructProperty.MapObjectSpawnerInStageSaveData.MapProperty.Value.StructProperty"),
+        String::from("StructProperty")
+    ), (
+        String::from("worldSaveData.StructProperty.MapObjectSpawnerInStageSaveData.MapProperty.Value.StructProperty.SpawnerDataMapByLevelObjectInstanceId.MapProperty.Key.StructProperty"),
+        String::from("Guid")
+    ), (
+        String::from("worldSaveData.StructProperty.MapObjectSpawnerInStageSaveData.MapProperty.Value.StructProperty.SpawnerDataMapByLevelObjectInstanceId.MapProperty.Value.StructProperty"),
+        String::from("StructProperty")
+    ), (
+        String::from("worldSaveData.StructProperty.MapObjectSpawnerInStageSaveData.MapProperty.Value.StructProperty.SpawnerDataMapByLevelObjectInstanceId.MapProperty.Value.StructProperty.ItemMap.MapProperty.Value.StructProperty"), 
+        String::from("StructProperty")
+    ), (
+        String::from("worldSaveData.StructProperty.BaseCampSaveData.MapProperty.Key.StructProperty"),
+        String::from("Guid")
+    ), (
+        String::from("worldSaveData.StructProperty.BaseCampSaveData.MapProperty.Value.StructProperty"),
+        String::from("StructProperty")
+    ), (
+        String::from("worldSaveData.StructProperty.BaseCampSaveData.MapProperty.Value.StructProperty.ModuleMap.MapProperty.Value.StructProperty"),
+        String::from("StructProperty")
+    ), (
+        String::from("worldSaveData.StructProperty.ItemContainerSaveData.MapProperty.Key.StructProperty"),
+        String::from("StructProperty")
+    ), (
+        String::from("worldSaveData.StructProperty.ItemContainerSaveData.MapProperty.Value.StructProperty"),
+        String::from("StructProperty")
+    ), (
+        String::from("worldSaveData.StructProperty.CharacterContainerSaveData.MapProperty.Key.StructProperty"),
+        String::from("StructProperty")
+    ), (
+        String::from("worldSaveData.StructProperty.CharacterContainerSaveData.MapProperty.Value.StructProperty"),
+        String::from("StructProperty")
+    ), (
+        String::from("worldSaveData.StructProperty.GroupSaveDataMap.MapProperty.Key.StructProperty"),
+        String::from("Guid")
+    ), (
+        String::from("worldSaveData.StructProperty.GroupSaveDataMap.MapProperty.Value.StructProperty"),
+        String::from("StructProperty")
+    ), (
+        String::from("worldSaveData.StructProperty.EnemyCampSaveData.StructProperty.EnemyCampStatusMap.MapProperty.Value.StructProperty"),
+        String::from("StructProperty")
+    )])
+}
